scripts/cli/requirements.txt
click>=8.1.0
requests>=2.31.0
rich>=13.0.0
python-dotenv>=1.0.0
pyinstaller>=6.0.0
